refactor: Move to Input namespace
